refactor(prj): separate tclient core code to core module
