+++ core/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/cenkalti/backoff/v4 v4.3.0
 	github.com/gabriel-vasile/mimetype v1.4.4
 	github.com/go-faster/errors v0.7.1
+	github.com/gotd/contrib v0.20.0
 	github.com/gotd/td v0.102.0
 	github.com/iyear/connectproxy v0.1.1
 	github.com/yapingcat/gomedia v0.0.0-20240601043430-920523f8e5c7
@@ -14,10 +15,12 @@ require (
 	go.uber.org/zap v1.27.0
 	golang.org/x/net v0.25.0
 	golang.org/x/sync v0.7.0
+	golang.org/x/time v0.5.0
 	gopkg.in/natefinch/lumberjack.v2 v2.2.1
 )
 
 require (
+	github.com/beevik/ntp v1.3.1 // indirect
 	github.com/go-faster/jx v1.1.0 // indirect
 	github.com/go-faster/xor v1.0.0 // indirect
 	github.com/gotd/ige v0.2.2 // indirect
